dtable code refactoring to support api model
