+++ ruleapi/dtable/expression.go
